drive__csv/requirements.txt
loguru==0.7.0
numpy==1.25.0
pandas==2.0.2
pip==23.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
setuptools==67.8.0
six==1.16.0
tzdata==2023.3
wheel==0.40.0
